Refactor project structure and update imports; add new RAG functionality and notebook for document processing
